API handles diary entry date and time separately
The usual timezone woes. Instead of a complete datetime string, we parse
the date and the time of day separately and put both of them together on
the server. As soon as code enters the application layer, all dates are
treated as UTC. This is likely going to be fine since a date in the diary
is nothing more than a named grouping mechanism.